REST-Service/requirements.txt
fastapi
uvicorn
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
ibm-cos-sdk==2.13.5
ibm-cos-sdk-core==2.13.5
ibm-cos-sdk-s3transfer==2.13.5
ibm_watson_machine_learning==1.0.359
ibm_watsonx_ai==1.0.10
idna==3.7
importlib_metadata==8.0.0
jmespath==1.0.1
joblib==1.4.2
lomond==0.3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.4
rouge==1.0.1
scikit-learn==1.5.0
scipy==1.14.0
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.1.0
zipp==3.19.2
openpyxl==3.1.5
langchain-ibm==0.1.10
celery==5.4.0
redis==5.0.7
flower==2.0.1
asyncio==3.4.3
python-dotenv
python-multipart
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
ibm-watsonx-gov==1.2.2
Jinja2==3.1.2
jsonschema==4.25.1
unitxt==1.26.6
textstat==0.7.10